Add logic for move movements
Implement the logic to make moves in a pure functional way. The state of
the game is kept in a Grid class instance. Over that instance we can
invoke the move method, passing the player and the position as
parameters. Depending on the current grid (game) state, that move can be
valid or not. If not valid, we indicate the reason. If valid, we return
a new Grid object with the updated state after that move.

Important note: the Grid objects are immutable; every time the move
method is successfully invoked a new instance of Grid is returned,
without any modification to the previous one. That immutability is
enforced by the compiler and very natural in Scala code.