Implement game logic using akka actors
And giving an actual implementation to the trait created in the previous
commit.

The use of akka actors is meant to deal with concurrency. We will have
in this server lots of concurrent games in progress between different
players. The state for all of those has to be maintained. Actors give us
a very easy to reason about programming model to keep that state in
memory without interferences between different games. When a game is
finished, it is persisted into MongoDB to allow future queries on it's
state. Also the leader board is updated appropriately.